requirements.txt
python-dotenv
requests
matplotlib
pillow
azure-ai-vision-imageanalysis==1.0.0b3
azure-cognitiveservices-vision-customvision==3.1.1
azure-ai-vision-face==1.0.0b2
azure-ai-textanalytics==5.3.0
azure-ai-language-questionanswering
azure-ai-language-conversations
azure-ai-translation-text==1.0.0b1